app/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
pandas==2.2.3
requests==2.32.3